+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/connectors/connectors.constants.js
@@ -1,0 +1,5 @@
+export const CONNECTOR_INTEGRATION_TYPE = 'kafkaConnect';
+
+export default {
+  CONNECTOR_INTEGRATION_TYPE,
+};