+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/general-information/general-information.component.js
@@ -23,6 +23,8 @@ export default {
     goToBackups: '<',
     goToMaintenances: '<',
     goToFork: '<',
+    goToIntegrations: '<',
+    goToSupportPage: '<',
     vRack: '<',
     vRackLink: '<',
     latestPlan: '<',