+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/general-information/general-information.routing.js
@@ -1,9 +1,7 @@
-import groupBy from 'lodash/groupBy';
 import find from 'lodash/find';
-import map from 'lodash/map';
 import { STATUS } from '../../../../../../components/project/storages/databases/databases.constants';
-import { INTEGRATION_TYPE } from '../../../../../../components/project/storages/databases/serviceIntegration.constants';
-import { DATABASE_TYPES, NODES_PER_ROW } from '../../databases.constants';
+import { NODES_PER_ROW } from '../../databases.constants';
+import { NEW_SUPPORT_TICKET_PARAMS } from './general-information.constants';
 import isFeatureActivated from '../../features.constants';
 
 export default /* @ngInject */ ($stateProvider) => {
@@ -52,6 +50,14 @@ export default /* @ngInject */ ($stateProvider) => {
           projectId,
           databaseId,
         }),
+      goToIntegrations: /* @ngInject */ ($state, databaseId, projectId) => () =>
+        $state.go(
+          'pci.projects.project.storages.databases.dashboard.service-integration',
+          {
+            projectId,
+            databaseId,
+          },
+        ),
       goToConfirmDeleteDatabase: /* @ngInject */ ($state, database) => (
         linkedServices,
       ) =>
@@ -106,6 +112,18 @@ export default /* @ngInject */ ($stateProvider) => {
             databaseId,
           },
         ),
+      goToSupportPage: /* @ngInject */ (coreURLBuilder, database) => () => {
+        const url = coreURLBuilder.buildURL(
+          'dedicated',
+          '/support/tickets/new',
+          {
+            categoryName: NEW_SUPPORT_TICKET_PARAMS.CATEGORY_NAME,
+            serviceTypeName:
+              NEW_SUPPORT_TICKET_PARAMS.BASE_SERVICE_TYPE + database.engine,
+          },
+        );
+        window.location.replace(url);
+      },
       goToFork: /* @ngInject */ ($state, database) => () =>
         $state.go('pci.projects.project.storages.databases.fork', {
           database,
@@ -236,47 +254,16 @@ export default /* @ngInject */ ($stateProvider) => {
         });
       },
       serviceIntegration: /* @ngInject */ (
-        $q,
         DatabaseService,
         database,
         projectId,
       ) =>
-        isFeatureActivated('showServiceIntegration', database.engine)
+        isFeatureActivated('serviceIntegrationTab', database.engine)
           ? DatabaseService.getIntegrations(
               projectId,
               database.engine,
               database.id,
-            ).then((integrations) => {
-              // filter integrations
-              const filter =
-                database.engine === DATABASE_TYPES.KAFKA
-                  ? [
-                      INTEGRATION_TYPE.MIRROR_MAKER,
-                      INTEGRATION_TYPE.KAFKA_CONNECT,
-                    ]
-                  : [INTEGRATION_TYPE.M3_AGGREGATOR];
-              const filteredIntegrations = integrations.filter((integration) =>
-                filter.includes(integration.type),
-              );
-              return $q
-                .all(
-                  // map servicename
-                  map(filteredIntegrations, (integration) => {
-                    return DatabaseService.getDatabaseDetails(
-                      projectId,
-                      integration.getTargetEngine(),
-                      integration.destinationServiceId,
-                    ).then((service) => {
-                      integration.setDestinationServiceName([service]);
-                      return integration;
-                    });
-                  }),
-                )
-                .then((aggregatedIntegrations) =>
-                  // group integrations by type
-                  groupBy(aggregatedIntegrations, 'type'),
-                );
-            })
+            )
           : [],
       users: /* @ngInject */ (DatabaseService, database, projectId) =>
         isFeatureActivated('usersTab', database.engine) ||