+++ packages/manager/modules/vps/src/dashboard/additional-disk/upgrade/upgrade.routing.js
@@ -1,0 +1,17 @@
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('vps.detail.dashboard.additional-disk-upgrade', {
+    url: '/additional-disk-upgrade',
+    views: {
+      modal: {
+        component: 'ovhManagerVpsComponentAdditionalDiskUpgrade',
+      },
+    },
+    layout: 'modal',
+    resolve: {
+      goBack: /* @ngInject */ ($state) => () => $state.go('^'),
+    },
+    atInternet: {
+      rename: 'vps::detail::additional-disk::upgrade',
+    },
+  });
+};